Refactors the test folder
