+++ tests/Rector/Namespace_/FactoryDefinitionRector/Fixture/Configured/without_args.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace RectorLaravel\Tests\Rector\Namespace_\FactoryDefinitionRector\Fixture;
+namespace RectorLaravel\Tests\Rector\Namespace_\FactoryDefinitionRector\Fixture\Configured;
 
 use RectorLaravel\Tests\Rector\Namespace_\FactoryDefinitionRector\Source\Model;
 
@@ -24,7 +24,7 @@ $factory->afterCreating(Model::class, function () {
 -----
 <?php
 
-namespace RectorLaravel\Tests\Rector\Namespace_\FactoryDefinitionRector\Fixture;
+namespace RectorLaravel\Tests\Rector\Namespace_\FactoryDefinitionRector\Fixture\Configured;
 
 use RectorLaravel\Tests\Rector\Namespace_\FactoryDefinitionRector\Source\Model;
 class ModelFactory extends \Illuminate\Database\Eloquent\Factories\Factory
