+++ tests/Rector/Namespace_/FactoryDefinitionRector/Fixture/Configured/without_namespace.php.inc
