+++ tests/Rector/Namespace_/FactoryDefinitionRector/Fixture/WithoutConfiguration/skip_reassign_factory.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace RectorLaravel\Tests\Rector\Namespace_\FactoryDefinitionRector\Fixture;
+namespace RectorLaravel\Tests\Rector\Namespace_\FactoryDefinitionRector\Fixture\WithoutConfiguration;
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
 $factory = 1;
@@ -12,7 +12,7 @@ $factory->define(\App\User::class, function (\Faker\Generator $faker) {
 -----
 <?php
 
-namespace RectorLaravel\Tests\Rector\Namespace_\FactoryDefinitionRector\Fixture;
+namespace RectorLaravel\Tests\Rector\Namespace_\FactoryDefinitionRector\Fixture\WithoutConfiguration;
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
 $factory = 1;
