+++ tests/Rector/Namespace_/FactoryDefinitionRector/Fixture/WithoutConfiguration/without_namespace.php.inc
