+++ tests/Rector/Namespace_/FactoryDefinitionRector/Source/AnotherModel.php
@@ -2,4 +2,4 @@
 
 namespace RectorLaravel\Tests\Rector\Namespace_\FactoryDefinitionRector\Source;
 
-class Model {}
+class AnotherModel {}
