+++ tests/Rector/Namespace_/FactoryDefinitionRector/config/configured_rule_without_configuration.php
