--- tests/Rector/Namespace_/FactoryDefinitionRectorConfiguredNoMatch/Fixture/skip_reassign_factory.php.inc
@@ -1,11 +1,0 @@
-<?php
-
-namespace RectorLaravel\Tests\Rector\Namespace_\FactoryDefinitionRector\Fixture;
-
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
-$factory = 1;
-$factory->define(\App\User::class, function (\Faker\Generator $faker) {
-    return [];
-});
-
-?>
