--- tests/Rector/Namespace_/FactoryDefinitionRectorConfiguredNoMatch/Fixture/without_namespace.php.inc
@@ -1,12 +1,0 @@
-<?php
-
-use App;
-
-$factory->define(RectorLaravel\Tests\Rector\Namespace_\FactoryDefinitionRector\Source\Model::class, function ($faker) {
-    return [
-        'name' => $faker->name,
-        'email' => $faker->unique()->safeEmail,
-    ];
-});
-
-?>
