--- tests/Rector/Namespace_/FactoryDefinitionRectorConfiguredNoMatch/config/configured_rule.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Rector\Config\RectorConfig;
-use RectorLaravel\Rector\Namespace_\FactoryDefinitionRector;
-
-return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->import(__DIR__ . '/../../../../../config/config.php');
-
-    $rectorConfig->ruleWithConfiguration(FactoryDefinitionRector::class, ['NonMatchingClass']);
-};
